src/requirements.txt
aiobotocore==2.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
boto3==1.39.11
botocore==1.39.11
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.62.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvclive==3.48.5
entrypoints==0.4
executing==2.2.0
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.8.0
hydra-core==1.3.2
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
jedi==0.19.2
jmespath==1.0.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kombu==5.5.4
markdown-it-py==4.0.0
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
nvidia-ml-py==12.575.51
omegaconf==2.3.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.7.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3fs==2025.7.0
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
stack-data==0.6.3
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
wrapt==1.17.3
yarl==1.20.1
zc.lockfile==3.0.post1
